Added addressing mode 1
